gallery add

+++ app/src/main/java/com/example/picgalleryapp/ui/camera/CameraViewModel.kt
@@ -1,42 +1,43 @@
 package com.example.picgalleryapp.ui.camera
 
 import android.content.Context
+import android.util.Log
 import androidx.databinding.ObservableField
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.bumptech.glide.Glide
-import com.davemorrissey.labs.subscaleview.ImageSource
-import com.example.picgalleryapp.data.models.Result
-import com.example.picgalleryapp.data.models.Result.Success
 import com.example.picgalleryapp.data.models.Result.Error
+import com.example.picgalleryapp.data.source.PicGalleryRepository
 import com.example.picgalleryapp.utils.SingleLiveEvent
 import com.example.picgalleryapp.utils.helpers.ImageHelper
 import com.otaliastudios.cameraview.PictureResult
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import java.io.File
 import java.lang.Exception
 
 /**
  * @author Tomislav Curis
  */
-class CameraViewModel(private val context : Context) : ViewModel(){
+class CameraViewModel(
+    private val context : Context,
+    private val repository: PicGalleryRepository) : ViewModel(){
 
     val isCameraVisible = ObservableField(true)
     val photo = ObservableField<File>()
 
     val takePhoto = SingleLiveEvent<Unit>()
+    val photoSaved = SingleLiveEvent<String>()
 
     fun photoTaken(image : PictureResult) {
 
         isCameraVisible.set(false)
 
         //https://youtrack.jetbrains.com/issue/IDEA-227359
-        viewModelScope.launch {
-
+        viewModelScope.launch(Dispatchers.IO) {
             try {
-                val file = getFile(image) as File
+                val file = Glide.with(context).downloadOnly().load(image.data).submit().get()
+                ImageHelper.resizeImage(file, 512)
                 photo.set(file)
                 isCameraVisible.set(false)
             } catch (e: Exception) {
@@ -44,20 +45,7 @@ class CameraViewModel(private val context : Context) : ViewModel(){
                 e.printStackTrace()
             }
         }
-
     }
-
-    suspend fun getFile(image : PictureResult) =
-        withContext(Dispatchers.IO) {
-           return@withContext try {
-                val file = Glide.with(context).downloadOnly().load(image.data).submit().get()
-                ImageHelper.resizeImage(file, 512)
-                file
-            } catch (e: Exception) {
-                Result.Error(e)
-            }
-        }
-
 
     fun takePhoto(){
         takePhoto.call()
@@ -65,24 +53,14 @@ class CameraViewModel(private val context : Context) : ViewModel(){
 
     fun saveRetake(save: Boolean) {
         isCameraVisible.set(!save)
-        photo.get()?.delete()
 
-
-    }
-
-
-    protected fun <T> handleResponseWithError(response: Result<T>): T? {
-        return when (response) {
-            is Result.Success -> {
-//                isDataLoadingError.value = false
-                response.data
+        if(save)
+            viewModelScope.launch {
+                repository.savePicture(photo.get().toString())
+                photoSaved.postValue("")
+                photo.set(null)
             }
-            is Result.Error -> {
-//                isDataLoadingError.value = true
-//                _error.postValue(response.exception)
-                null
-            }
-            is Result.Loading -> null
-        }
+        else
+            photo.get()?.delete()
     }
 }
